+++ eras/dijkstra/impl/cddl/exe/GenerateCBOR.hs
@@ -1,7 +1,0 @@
-module Main where
-
-import Cardano.Ledger.Dijkstra.HuddleSpec (dijkstraCDDL)
-import Test.Cardano.Ledger.Binary.Cuddle.GenerateCBOR (generateCBORMain)
-
-main :: IO ()
-main = generateCBORMain dijkstraCDDL