+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/Cuddle/GenerateCBOR.hs
@@ -19,6 +19,9 @@ import Control.Monad (forM_, when)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Base16 as Base16
 import qualified Data.ByteString.Char8 as BS8
+import Data.List (intercalate)
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
 import Data.Maybe (fromMaybe, isJust)
 import qualified Data.Text as T
 import Options.Applicative ((<**>))
@@ -32,17 +35,23 @@ import Test.QuickCheck.Gen (unGen)
 import Test.QuickCheck.Random (mkQCGen)
 
 data GenerateCBOROpts = GenerateCBOROpts
-  { gcboRuleNames :: ![T.Text]
+  { gcboEra :: !T.Text
+  , gcboRuleNames :: ![T.Text]
   , gcboZap :: !(Maybe Int)
   , gcboCount :: !Int
   , gcboSeed :: !(Maybe Int)
   , gcboBinary :: !Bool
   }
 
-optsParser :: Opt.Parser GenerateCBOROpts
-optsParser =
+optsParser :: [String] -> Opt.Parser GenerateCBOROpts
+optsParser eras =
   GenerateCBOROpts
-    <$> Opt.some
+    <$> Opt.strOption
+      ( Opt.long "era"
+          <> Opt.metavar "ERA"
+          <> Opt.help ("Era to generate CBOR for. One of: " <> intercalate ", " eras)
+      )
+    <*> Opt.some
       ( Opt.strArgument $
           Opt.metavar "RULE_NAME..."
             <> Opt.help "CDDL rule names to generate CBOR for"
@@ -73,16 +82,24 @@ optsParser =
           <> Opt.help "Output raw CBOR bytes instead of hex encoding"
       )
 
-generateCBORMain :: Cuddle.Huddle -> IO ()
-generateCBORMain huddle = do
+generateCBORMain :: Map T.Text Cuddle.Huddle -> IO ()
+generateCBORMain eraCDDLs = do
   opts <-
     Opt.execParser $
       Opt.info
-        (optsParser <**> Opt.helper)
+        (optsParser eraNames <**> Opt.helper)
         ( Opt.fullDesc
             <> Opt.progDesc "Generate CBOR data from Cardano Ledger CDDL rules"
             <> Opt.header "generate-cbor - CBOR data generator from Cardano Ledger CDDL specifications"
         )
+  huddle <- case Map.lookup (gcboEra opts) eraCDDLs of
+    Nothing ->
+      die $
+        "Unknown era: "
+          <> T.unpack (gcboEra opts)
+          <> "\nSupported eras: "
+          <> intercalate ", " eraNames
+    Just h -> pure h
   case resolveHuddle huddle of
     Left err -> die $ "Failed to resolve CDDL: " <> err
     Right root -> do
@@ -94,6 +111,8 @@ generateCBORMain huddle = do
           hPutStrLn stderr $
             "# " <> T.unpack ruleName
         forM_ [0 .. gcboCount opts - 1] $ emitSample opts env ruleName
+  where
+    eraNames = map T.unpack (Map.keys eraCDDLs)
 
 emitSample :: GenerateCBOROpts -> HuddleEnv -> T.Text -> Int -> IO ()
 emitSample opts env ruleName sampleIx = do